OSIU_repo_scaffold/requirements.txt
pandas>=2.2.0
numpy>=1.26.0
requests>=2.32.0
PyYAML>=6.0.1
matplotlib>=3.8.0
tabulate>=0.9.0